Core/Misc: money fixes.

+++ src/server/game/Entities/Pet/Pet.cpp
@@ -1654,7 +1654,7 @@ bool Pet::resetTalents (bool no_cost)
         return false;
     }
 
-    uint32 cost = 0;
+    uint64 cost = 0;
 
     if (!no_cost)
     {
