+++ src/server/game/Handlers/AuctionHouseHandler.cpp
@@ -221,7 +221,7 @@ void WorldSession::HandleAuctionSellItem (WorldPacket & recv_data)
     AuctionHouseObject* auctionHouse = sAuctionMgr->GetAuctionsMap(pCreature->getFaction());
 
     //we have to take deposit :
-    uint32 deposit = sAuctionMgr->GetAuctionDeposit(auctionHouseEntry, etime, it, count);
+    uint64 deposit = sAuctionMgr->GetAuctionDeposit(auctionHouseEntry, etime, it, count);
     if (!pl->HasEnoughMoney(deposit))
     {
         SendAuctionCommandResult(0, AUCTION_SELL_ITEM, AUCTION_NOT_ENOUGHT_MONEY);
@@ -327,7 +327,7 @@ void WorldSession::HandleAuctionPlaceBid (WorldPacket & recv_data)
         return;
     }
 
-    if (!pl->HasEnoughMoney((uint32) price))
+    if (!pl->HasEnoughMoney((uint64) price))
     {
         //you don't have enought money!, client tests!
         //SendAuctionCommandResult(auction->auctionId, AUCTION_PLACE_BID, ???);
@@ -425,7 +425,7 @@ void WorldSession::HandleAuctionRemoveItem (WorldPacket & recv_data)
         {
             if (auction->bidder > 0)          // If we have a bidder, we have to send him the money he paid
             {
-                uint32 auctionCut = auction->GetAuctionCut();
+                uint64 auctionCut = auction->GetAuctionCut();
                 if (!pl->HasEnoughMoney(auctionCut))          //player doesn't have enough money, maybe message needed
                     return;
                 //some auctionBidderNotification would be needed, but don't know that parts..
